use erfa float

+++ src/astro.h
@@ -5,16 +5,12 @@
 #include <time.h>
 
 typedef struct {
-    double x;
-    double y;
-    double intensity;
+    float x;
+    float y;
+    float intensity;
 } star_t;
 
 typedef void (*star_callback_t)(star_t const*);
 
 void get_stars(
-  struct tm const* time,
-  double subsecond,
-  double latitude,
-  double longitude,
-  star_callback_t on_star);
+  struct tm const* time, float subsecond, float latitude, float longitude, star_callback_t on_star);
